requirements.txt
dash
dash-bootstrap-components
pandas
yfinance
plotly
redis
os
python-dotenv
zlib
pickle